fix(scripts): address iteration-2 review blockers for #1017
- Extract mode functions to scripts/_modes.py (gen_nkeys.py → 296 lines, under 300 LOC limit)
- Fix _get_provider() silent fake fallback: call ensure_nk_or_exit() instead
- Fix 6 stale '(not yet implemented)' help strings across genkeys flags
- Deduplicate _cmd_validate_supervisor vs --validate-supervisor branch
- Remove dead _fake_pubkeys() and _NOT_YET constant
- Add # verified: annotations to all 10 TestLoadMatrixNegatives tests
- Add voice-tts + clipool-worker live connect tests to test_parity_e2e.py
- Add retired identity connection rejection test (separate nats-server instance)
- Remove scripts/gen_nkeys.py from tools/file_exemptions.txt
